Travel List React web app
Travel List is a react web app which helps users to add new items, mark as packed, delete items and also can clear the list for the travel tour. There is a sorting filter to sort based on description, packed items and by input. It also shows how much items are packed and are remaining from the total items.